% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem.R
\name{ABSLOPE}
\alias{ABSLOPE}
\title{Create an adaSlope model with full algorithm}
\usage{
ABSLOPE(
  X,
  y,
  lambda,
  a,
  b,
  beta.start = NA,
  maxit = 300,
  print_iter = FALSE,
  tol_em = 1e-06,
  impute = "PCA",
  sigma.known = NA,
  sigma.init = NA,
  scale = FALSE,
  method_na = "lineq"
)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response vector}

\item{lambda}{Vector of coefficient L1 penalties}

\item{a}{Prior for coefficient calculation}

\item{b}{Prior for coefficient calculation}

\item{print_iter}{}
}
\value{
A SLOPE model
}
\description{
Create an adaSlope model with full algorithm
}
\examples{
X = matrix(rnorm(1000), nrow=100)
b = c(sample(-5:5, 5), rep(0, 5))
y = X \%*\% b + rnorm(100, 0, 0.1)
A <- ABSLOPE(X, y,lambda=seq(10, 5, length.out=10), a=1, b=1)
}
