% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slobe.R
\name{SLOBE}
\alias{SLOBE}
\title{Create an adaSlope model with simplified algorithm}
\usage{
SLOBE(
  X,
  y,
  lambda,
  a = 1,
  b = 1,
  beta.start = NA,
  maxit = 500,
  tol_em = 1e-05,
  impute = "PCA",
  sigma.known = NA,
  sigma.init = NA,
  print_iter = FALSE
)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response vecor}

\item{lambda}{Vector of coefficient L1 penalties}

\item{a}{Prior for coefficient calculation}

\item{b}{Prior for coefficient calculation}

\item{print_iter}{}
}
\value{
A SLOPE model
}
\description{
Create an adaSlope model with simplified algorithm
}
\examples{
X = matrix(rnorm(1000), nrow=100)
b = c(sample(-5:5, 5), rep(0, 5))
y = X \%*\% b + rnorm(100, 0, 0.1)
A <- SLOBE(X, y, lambda=seq(10, 5, length.out=10))
}
